fix issues

+++ bpf_helpers/trace2.bpf.c
@@ -12,7 +12,7 @@ int bpf_prog1(struct syscall_trace_enter* ctx)
 		return 0;
 	}
     int len=0;
-    for(long unsigned i=0;i<1<<10;i++)
+    for(int i=0;i<1<<10;i++)
     {
         if(fname[i]!='\0')
             len++;
